requirements.txt
datasets
pandas
numpy
scikit-learn
transformers
accelerate
evaluate
matplotlib
pyyaml
fasttext-langdetect
regex
tqdm
joblib
